+++ accounting/src/apideck_accounting_unify/models/accounting_ledgeraccountsaddop.py
@@ -6,7 +6,7 @@ from .createledgeraccountresponse import (
     CreateLedgerAccountResponseTypedDict,
 )
 from .httpmetadata import HTTPMetadata, HTTPMetadataTypedDict
-from .ledgeraccount import LedgerAccountInput, LedgerAccountInputTypedDict
+from .ledgeraccount_input import LedgerAccountInput, LedgerAccountInputTypedDict
 from .unexpectederrorresponse import (
     UnexpectedErrorResponse,
     UnexpectedErrorResponseTypedDict,