+++ accounting/src/apideck_accounting_unify/models/invoice.py
@@ -10,7 +10,7 @@ from .deprecatedlinkedtrackingcategory import (
     DeprecatedLinkedTrackingCategory,
     DeprecatedLinkedTrackingCategoryTypedDict,
 )
-from .invoicelineitem import (
+from .invoicelineitem_input import (
     InvoiceLineItem,
     InvoiceLineItemInput,
     InvoiceLineItemInputTypedDict,