+++ accounting/src/apideck_accounting_unify/subsidiaries.py
@@ -91,7 +91,7 @@ class Subsidiaries(BaseSDK):
 
         retry_config = None
         if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["5XX"])
+            retry_config = (retries, ["408", "500", "502", "503", "504"])
 
         http_res = self.do_request(
             hook_ctx=HookContext(
@@ -272,7 +272,7 @@ class Subsidiaries(BaseSDK):
 
         retry_config = None
         if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["5XX"])
+            retry_config = (retries, ["408", "500", "502", "503", "504"])
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
@@ -468,7 +468,7 @@ class Subsidiaries(BaseSDK):
 
         retry_config = None
         if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["5XX"])
+            retry_config = (retries, ["408", "500", "502", "503", "504"])
 
         http_res = self.do_request(
             hook_ctx=HookContext(
@@ -640,7 +640,7 @@ class Subsidiaries(BaseSDK):
 
         retry_config = None
         if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["5XX"])
+            retry_config = (retries, ["408", "500", "502", "503", "504"])
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
@@ -794,7 +794,7 @@ class Subsidiaries(BaseSDK):
 
         retry_config = None
         if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["5XX"])
+            retry_config = (retries, ["408", "500", "502", "503", "504"])
 
         http_res = self.do_request(
             hook_ctx=HookContext(
@@ -948,7 +948,7 @@ class Subsidiaries(BaseSDK):
 
         retry_config = None
         if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["5XX"])
+            retry_config = (retries, ["408", "500", "502", "503", "504"])
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
@@ -1123,7 +1123,7 @@ class Subsidiaries(BaseSDK):
 
         retry_config = None
         if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["5XX"])
+            retry_config = (retries, ["408", "500", "502", "503", "504"])
 
         http_res = self.do_request(
             hook_ctx=HookContext(
@@ -1298,7 +1298,7 @@ class Subsidiaries(BaseSDK):
 
         retry_config = None
         if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["5XX"])
+            retry_config = (retries, ["408", "500", "502", "503", "504"])
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
@@ -1449,7 +1449,7 @@ class Subsidiaries(BaseSDK):
 
         retry_config = None
         if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["5XX"])
+            retry_config = (retries, ["408", "500", "502", "503", "504"])
 
         http_res = self.do_request(
             hook_ctx=HookContext(
@@ -1600,7 +1600,7 @@ class Subsidiaries(BaseSDK):
 
         retry_config = None
         if isinstance(retries, utils.RetryConfig):
-            retry_config = (retries, ["5XX"])
+            retry_config = (retries, ["408", "500", "502", "503", "504"])
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(