+++ accounting/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [python v0.11.0] accounting
 ### Releases
 - [PyPI v0.11.0] https://pypi.org/project/apideck-accounting-unify/0.11.0 - accounting
+
+## 2025-04-08 08:33:06
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.531.0 (2.568.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.11.1] accounting
+### Releases
+- [PyPI v0.11.1] https://pypi.org/project/apideck-accounting-unify/0.11.1 - accounting